feat: add map for find all wines response DTO

+++ src/main/java/com/example/menu/wines/WinesRepository.java
@@ -2,5 +2,5 @@ package com.example.menu.wines;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WinesRepository extends JpaRepository<Wine, Long> {
+public interface WinesRepository extends JpaRepository<Wine, Integer> {
 }
